# admin-be/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Admin Backend",
        "description": "Admin Backend Description",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://admin-apc.dotbeit.com",
            "description": "Admin Backend"
        }
    ],
    "paths": {
...
        "/api/clinics": {
            "get": {
                "tags": [
                    "Clinic Pages"
                ],
                "operationId": "clinics",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "query",
                        "description": "Lang",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "ar"
                        }
                    },
                    {
                        "name": "query_search",
                        "in": "query",
                        "description": "search in the clinics",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "with_specialities",
                        "in": "query",
                        "description": "to search in the clinics and the specialities as well",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "clinics",
                        "in": "query",
                        "description": "the clinics",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Cardia,test"
                    },
                    {
                        "name": "specialities",
                        "in": "query",
                        "description": "the specialities",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "heart"
                    },
                    {
                        "name": "gender",
                        "in": "query",
                        "description": "the gender",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "male"
                    },
                    {
                        "name": "all_doctors",
                        "in": "query",
                        "description": "get all doctors",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        },
                        "example": "true"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "For Home Data as ['clinics']"
                    }
                }
            }
        },
...
        "/api/pages": {
            "get": {
                "tags": [
                    "Pages"
                ],
                "operationId": "clinics",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "query",
                        "description": "Lang",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "ar"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "For Home Data as ['pages']"
                    }
                }
            }
        },
        "/api/Page/{slug}": {
            "get": {
                "tags": [
                    "Pages"
                ],
                "operationId": "page",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "query",
                        "description": "Lang",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "ar"
                        }
                    },
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "hello-world"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "For  Data as ['page']"
                    },
                    "404": {
                        "description": "Slug Not Found"
                    }
                }
            }
        },
...
        "/api/category/{slug}": {
            "get": {
                "tags": [
                    "Post Pages"
                ],
                "operationId": "category",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "query",
                        "description": "Lang",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "ar"
                        }
                    },
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "category-1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "For Home Data as ['Category']"
                    },
                    "404": {
                        "description": "Slug Not Found"
                    }
                }
            }
        },
...
        "/api/service/{slug}": {
            "get": {
                "tags": [
                    "Service Pages"
                ],
                "operationId": "service",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "query",
                        "description": "Lang",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "ar"
                        }
                    },
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "وحدة-قسطرة-القلب"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "For Home Data as ['service']"
                    },
                    "404": {
                        "description": "Slug Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}